Extract portable SIMD functionality into rten-simd crate
Extract the portable SIMD types, functions and dispatch logic out of
rten-vecmath and into a new rten-simd crate.

In the process the dispatch logic was refactored to avoid duplicating the logic
to decide which ISA to use for a particular operation. There is now a
`SimdDispatcher` struct which handles dispatching an operation that implements a
`SimdOp` trait. `SimdOp` has a single `eval` method that takes a generic
argument specifying the concrete `SimdFloat` impl, but no arguments.
