+++ Cargo.toml
@@ -4,8 +4,10 @@ members = [
   "rten-cli",
   "rten-imageio",
   "rten-imageproc",
+  "rten-simd",
   "rten-tensor",
   "rten-text",
+  "rten-vecmath",
 
   # Development crates. These are not published.
   "rten-bench",
@@ -39,6 +41,7 @@ rayon = "1.7.0"
 smallvec = { version = "1.10.0", features = ["union", "const_generics", "const_new"] }
 rten-tensor = { path = "./rten-tensor", version = "0.9.0" }
 rten-vecmath = { path = "./rten-vecmath", version = "0.9.0" }
+rten-simd = { path = "./rten-simd", version = "0.9.0" }
 fastrand = { version = "2.0.2", optional = true }
 fastrand-contrib = { version = "0.1.0", optional = true }
 rustc-hash = "1.1.0"
@@ -55,7 +58,7 @@ crate-type = ["lib", "cdylib"]
 
 [features]
 # Use AVX-512 instructions if available. Requires nightly Rust for AVX-512 intrinsics.
-avx512 = ["rten-vecmath/avx512"]
+avx512 = ["rten-simd/avx512", "rten-vecmath/avx512"]
 # Enable loading models using memory mapping
 mmap = ["memmap2"]
 # Generate WebAssembly API using wasm-bindgen.
