+++ src/gemm/kernels/wasm.rs
@@ -1,9 +1,9 @@
 use std::mem::MaybeUninit;
 use std::ops::Range;
 
+use rten_simd::arch::wasm::v128f;
+use rten_simd::SimdFloat;
 use rten_tensor::Matrix;
-use rten_vecmath::simd_vec::wasm::v128f;
-use rten_vecmath::simd_vec::SimdFloat;
 
 use super::{simd_gemm, simd_gemv, Kernel};
 use crate::gemm::packing::{pack_a_block, pack_b_block};
